TimeScale Slider Fix

+++ Assets/Scripts/OrbitControl.cs
@@ -20,9 +20,9 @@ public class OrbitControl : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        float t = Constants.CurrentWorldRealTime % PlanetData.Period;
-        float orbitOffset = t / PlanetData.Period;
-        float orbitPosition = PlanetData.StartOffset + orbitOffset * Mathf.PI * 2;
+        double t = Constants.CurrentWorldRealTime % PlanetData.Period;
+        double orbitOffset = t / PlanetData.Period;
+        float orbitPosition = (float) (PlanetData.StartOffset + orbitOffset * Mathf.PI * 2);
         float x = PlanetData.OrbitRadius * Mathf.Cos(orbitPosition);
         float z = PlanetData.OrbitRadius * Mathf.Sin(orbitPosition);
         gameObject.transform.localPosition = new Vector3(x, 0, z);
